added back create event modal and added confirmation for when an event is going to be deleted

+++ src/components/AdminPage.vue
@@ -46,6 +46,8 @@
 
     </modal>
 
+    <CreateEventModal></CreateEventModal>
+
 
     <br />
     <br />
@@ -99,6 +101,7 @@ import NavigationBar from './NavigationBar'
 import Footer from './Footer'
 import VueBootstrap4Table from 'vue-bootstrap4-table'
 import AuthenticationService from '../services/AuthenticationService'
+import CreateEventModal from '../modals/CreateEventModal'
 
 export default {
   name: 'AdminPage',
@@ -205,6 +208,7 @@ export default {
     NavigationBar,
     Footer,
     VueBootstrap4Table,
+    CreateEventModal
   },
   methods: {
     openModal (row) {
@@ -228,11 +232,16 @@ export default {
     async deleteEvent (row) {
       // eslint-disable-next-line
       console.log(row);
-      // let del
-      // alert("Are you sure you want to delete this event?")
-      await AuthenticationService.deleteEvent({
-        data: row
-      })
+
+      let del = confirm("Are you sure you want to delete this event?")
+
+      if (del) {
+        await AuthenticationService.deleteEvent({
+          data: row
+        })
+      } else {
+        return false
+      }
     }
   },
   async created () {
